requirements.txt
anthropic
fastapi
mangum
python-dotenv